requirements.txt
requests
beautifulsoup4
faiss-cpu
sentence-transformers
langchain
langchain-openai
urllib3
python-dotenv


